correct directory name

+++ PracticalApps/Inje.AIConvergence.Common.EntityModels.Sqlite/Category.cs
@@ -15,7 +15,7 @@ namespace AIConvergence.Shared
     }
 
     [Key]
-    public long CategoryId { get; set; }
+    public int CategoryId { get; set; }
     [Required]
     [Column(TypeName = "nvarchar (15)")]
     [StringLength(15)]
