+++ PracticalApps/Inje.AIConvergence.Common.EntityModels.Sqlite/OrderDetail.cs
@@ -15,14 +15,14 @@ namespace AIConvergence.Shared
     {
         [Key]
         [Column(TypeName = "INT")]
-        public long OrderId { get; set; }
+        public int OrderId { get; set; }
         [Key]
         [Column(TypeName = "INT")]
-        public long ProductId { get; set; }
+        public int ProductId { get; set; }
         [Column(TypeName = "money")]
         public decimal? UnitPrice { get; set; } = null!;
         [Column(TypeName = "smallint")]
-        public long Quantity { get; set; }
+        public int Quantity { get; set; }
         public double Discount { get; set; }
 
         [ForeignKey("OrderId")]
